Move job application models to separate module
