+++ src/repository.rs
@@ -1,1 +1,2 @@
 pub mod job_application_repository;
+pub mod job_application_model;
